fixed / updated nk65 keymap, so it works again and is tested
QMK replaced this board's Wilbatech RGB with RGB Matrix,
so it uses the same code as everything else now.
This is easier to maintain,
though it does have the downside of no more pretty static gradient mode.
However, instead of that, it now has the multisplash effect I like.

The indicator LEDs are also easier now, since they're treated as
just regular LEDs in the RGB Matrix.